Moved the validation to configurator
